Change sleep timer to toggle power switch at end of sleep interval
Currently, blocking is turned on when the sleep timer starts and off
when the timer expires. Instead, retain the current state when the timer
starts and change to the other state when the timer expires. That way,
I can set a timer to read news and be blocked from it when the timer
expires and I need to do work.